add debug option in the to_verilog function; refactor the code
