+++ magma/circuit.py
@@ -23,7 +23,8 @@ from magma.syntax.combinational import combinational
 from magma.syntax.sequential import sequential
 from magma.syntax.combinational import combinational
 from magma.syntax.sequential import sequential
-from magma.syntax.verilog import combinational_to_verilog, sequential_to_verilog
+from magma.syntax.verilog import combinational_to_verilog, \
+    sequential_to_verilog, build_kratos_debug_info
 if sys.version_info > (3, 0):
     from functools import reduce
